fix: created Favoritecomponent
